+++ src/commons/constants.py
@@ -72,6 +72,7 @@ OUTPUT_STORAGE_ATTR = 'output_storage'
 ADDED_AT_ATTR = 'added_at'
 RECOMMENDATION_TYPE_ATTR = 'recommendation_type'
 JOB_ID_ATTR = 'job_id'
+LIMIT_ATTR = 'limit'
 
 RULE_ACTION_ATTR = 'rule_action'
 CONDITION_ATTR = 'condition'
@@ -260,6 +261,7 @@ ENV_VAULT_HOST = 'VAULT_URL'
 ENV_VAULT_PORT = 'VAULT_SERVICE_SERVICE_PORT'
 
 ENV_TENANT_CUSTOMER_INDEX = 'tenants_customer_name_index_rcu'
+ENV_LM_TOKEN_LIFETIME_MINUTES = 'lm_token_lifetime_minutes'
 EXP_ATTR = 'exp'
 
 # cognito
@@ -305,3 +307,4 @@ B64ENCODED_ATTR = 'b64_encoded'
 ALLOWED_ATTR = 'allowed'
 FORBIDDEN_ATTR = 'forbidden'
 REMAINING_BALANCE_ATTR = 'remaining_balance'
+TOKEN_ATTR = 'token'