+++ src/lambdas/r8s_api_handler/processors/parent_licenses_processor.py
@@ -361,7 +361,9 @@ class ParentLicensesProcessor(AbstractCommandProcessor):
     def _execute_license_sync(self, license_obj: License, customer: str):
         _LOG.info(f'Syncing license \'{license_obj.license_key}\'')
         response = self.license_manager_service.synchronize_license(
-            license_key=license_obj.license_key)
+            license_key=license_obj.license_key,
+            customer=customer
+        )
         if not response.status_code == 200:
             return