+++ r8s/CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.19.10] - 2023-11-22
+* add `limit` parameter to `r8s job describe` command
+
 ## [2.15.0] - 2023-09-29
 * Upgrade r8s CLI dependency versions to support python 3.10